+++ docs/RELEASE-NOTES.md
@@ -4,6 +4,7 @@ These are the release notes for the [Maps extension](../README.md).
 
 * Re added Google Earth support
 * Removed support for the deprecated Google JavaScript API
+* Updated the translations to use the new MediaWiki JSON format
 
 ## Maps 3.0.1